feat(datastore): cost the stored-schema cache by estimated byte size
The stored-schema cache previously admitted every entry at a fixed cost of 1, so a
MaxCost configured in bytes (e.g. the server's 32MiB default, or the embedded
SchemaCacheMaxCostBytes) actually bounded the number of cached schema versions rather
than memory.

Cost each entry by ReadOnlyStoredSchema.EstimatedSize() instead: a rough schema byte
size plus, for each registered schema-derived cache kind, an estimate of the bytes it
adds once populated. Datastore readers pass the exact serialized size they just read;
other callers fall back to a cheap schema-text proxy, avoiding a full proto walk
(proto.Size is ~1.5ms on a 1k-definition schema). RegisterDerivedCache now takes a size
estimator, and the compiled-caveat cache estimates per-caveat CEL overhead.

+++ pkg/datalayer/hashcache_test.go
@@ -59,8 +59,8 @@ func TestSchemaHashCache_SetCostsBySize(t *testing.T) {
 	require.NoError(t, shc.Set(SchemaHash("small"), small))
 	require.NoError(t, shc.Set(SchemaHash("large"), large))
 
-	require.Equal(t, int64(small.EstimatedSize()), tc.cost(SchemaCacheKey("small")))
-	require.Equal(t, int64(large.EstimatedSize()), tc.cost(SchemaCacheKey("large")))
+	require.Equal(t, small.EstimatedSize(), tc.cost(SchemaCacheKey("small")))
+	require.Equal(t, large.EstimatedSize(), tc.cost(SchemaCacheKey("large")))
 	require.Greater(t, tc.cost(SchemaCacheKey("large")), tc.cost(SchemaCacheKey("small")))
 	require.Greater(t, tc.cost(SchemaCacheKey("small")), int64(1))
 }
